Store & overwrite catalog internally
Remove selector_catalog_filepath parameter and use a
temp_catalog.json as the output of the discover &
selection commands. After execution of each of these
commands, the temp_catalog.json is used to overwrite
the tap_catalog_filepath. This prevents issues when the
same catalog name is used as both a command parameter
(i.e. --catalog catalog.json) and an output parameter
(i.e. > catalog.json) which throws a JSONDecoder error.